divide main file in to sub files
Its easy to read and manage app
