+++ lib/components/mini_player.dart
@@ -1,0 +1,72 @@
+// components/mini_player.dart - Mini player shown in playlist view
+import 'package:flutter/material.dart';
+import '../models/song.dart';
+
+class MiniPlayer extends StatelessWidget {
+  final Song currentSong;
+  final bool isPlaying;
+  final VoidCallback onTogglePlayPause;
+
+  const MiniPlayer({
+    super.key,
+    required this.currentSong,
+    required this.isPlaying,
+    required this.onTogglePlayPause,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.black26,
+      padding: const EdgeInsets.symmetric(
+        vertical: 8.0,
+        horizontal: 16.0,
+      ),
+      child: Row(
+        children: [
+          ClipRRect(
+            borderRadius: BorderRadius.circular(4),
+            child: Image.asset(
+              currentSong.imageUrl,
+              width: 36,
+              height: 36,
+              fit: BoxFit.cover,
+            ),
+          ),
+          const SizedBox(width: 12),
+          Expanded(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                Text(
+                  currentSong.title,
+                  style: const TextStyle(
+                    color: Colors.white,
+                    fontWeight: FontWeight.bold,
+                  ),
+                  maxLines: 1,
+                  overflow: TextOverflow.ellipsis,
+                ),
+                Text(
+                  currentSong.artist,
+                  style: TextStyle(
+                    color: Colors.white.withOpacity(0.7),
+                    fontSize: 12,
+                  ),
+                ),
+              ],
+            ),
+          ),
+          IconButton(
+            onPressed: onTogglePlayPause,
+            icon: Icon(
+              isPlaying ? Icons.pause : Icons.play_arrow,
+              color: Colors.white,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
