+++ lib/components/progress_bar.dart
@@ -1,0 +1,71 @@
+// components/progress_bar.dart - Audio progress bar component
+import 'package:flutter/material.dart';
+import '../utils/audio_helpers.dart';
+
+class ProgressBar extends StatelessWidget {
+  final Duration position;
+  final Duration duration;
+  final Function(double) onChanged;
+
+  const ProgressBar({
+    super.key,
+    required this.position,
+    required this.duration,
+    required this.onChanged,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 24.0),
+      child: Column(
+        children: [
+          SliderTheme(
+            data: SliderTheme.of(context).copyWith(
+              trackHeight: 4.0,
+              thumbShape: const RoundSliderThumbShape(
+                enabledThumbRadius: 6.0,
+              ),
+              overlayShape: const RoundSliderOverlayShape(
+                overlayRadius: 14.0,
+              ),
+              activeTrackColor: Colors.white,
+              inactiveTrackColor: Colors.white.withOpacity(0.2),
+              thumbColor: Colors.white,
+              overlayColor: Colors.white.withOpacity(0.2),
+            ),
+            child: Slider(
+              value: position.inSeconds.toDouble(),
+              max: duration.inSeconds.toDouble() > 0
+                  ? duration.inSeconds.toDouble()
+                  : 1.0,
+              onChanged: onChanged,
+            ),
+          ),
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 16.0),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Text(
+                  AudioHelpers.formatTime(position),
+                  style: TextStyle(
+                    color: Colors.white.withOpacity(0.7),
+                    fontSize: 14,
+                  ),
+                ),
+                Text(
+                  AudioHelpers.formatTime(duration),
+                  style: TextStyle(
+                    color: Colors.white.withOpacity(0.7),
+                    fontSize: 14,
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
